nbits & nbits_from_target
for specific coins we should submit block to coin daemon
not when the hash < target(nbits), but also when hash < target,
received from daemon. mean, target = getblocktemplate target,
not target from nbits. sorry, may be not clear at all, but
i'm very tired. so, let this value be commented for now.